requirements.txt
streamlit
python-dotenv
PyPDF2
google-generativeai
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langchain-huggingface
numpy
# Firebase & Visuals
firebase-admin
pyrebase4
pyvis
# RAG Core (FAISS)
faiss-cpu
sentence-transformers
tiktoken
pypdf
fpdf2
python-docx
youtube-transcript-api
yt-dlp
